+++ src/lib/Scenes/Partner/Components/PartnerMap.tsx
@@ -2,7 +2,7 @@ import { useActionSheet } from "@expo/react-native-action-sheet"
 import MapboxGL from "@react-native-mapbox-gl/maps"
 import { themeGet } from "@styled-system/theme-get"
 import { PartnerMap_location } from "__generated__/PartnerMap_location.graphql"
-import { cityAndPostalCode, tappedOnMap } from "lib/Components/LocationMap"
+import { cityAndPostalCode, tappedOnMap } from "lib/Components/LocationMap/LocationMap"
 import { Pin } from "lib/Icons/Pin"
 import { ArtsyMapStyleURL } from "lib/Scenes/Map/GlobalMap"
 import { Box, Flex, Sans, Serif, Spacer } from "palette"
