+++ src/lib/Scenes/Show/Components/ShowLocation.tsx
@@ -1,5 +1,5 @@
 import { ShowLocation_show } from "__generated__/ShowLocation_show.graphql"
-import { LocationMapContainer as LocationMap } from "lib/Components/LocationMap"
+import { LocationMapContainer as LocationMap } from "lib/Components/LocationMap/LocationMap"
 import { Box, BoxProps } from "palette"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
